Split LitUI and UnlitUI with base class to share behavior with the MaterialGraph
