+++ Assets/ScriptableRenderLoop/HDRenderLoop/Material/Lit/Editor/BaseLitUI.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d35c4b8ec5d49e4438d0b0ba2655c495
+timeCreated: 1480349253
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
