+++ Assets/ScriptableRenderLoop/HDRenderLoop/Material/Unlit/Editor/BaseUnlitUI.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 11fd029bb88b2c04092b1e4d71b919c6
+timeCreated: 1480351138
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
