+++ Assets/ScriptableRenderLoop/HDRenderLoop/Material/Unlit/Unlit.shader
@@ -246,5 +246,5 @@ Shader "HDRenderLoop/Unlit"
         }
     }
 
-    CustomEditor "UnlitGUI"
+    CustomEditor "Experimental.ScriptableRenderLoop.UnlitGUI"
 }
